core_service_mapped/pyproject.toml
[project]
name = "core_service_mapped"
version = "1.0.0"
requires-python = ">=3.10"
dependencies = [
  "fastapi==0.115.0",
  "uvicorn[standard]==0.30.6",
  "pydantic==2.9.2",
  "pydantic-settings==2.5.2",
  "python-jose==3.3.0",
  "passlib[bcrypt]==1.7.4",
  "SQLAlchemy==2.0.36",
  "psycopg2-binary==2.9.9",
  "httpx==0.27.2",
  "jinja2==3.1.4",
  "prometheus-client==0.21.0",
  "python-multipart==0.0.12",
]

[tool.uvicorn]
host = "0.0.0.0"
port = 8000
reload = true
